enhance background image to one hour
